# specs/001-build-a-web/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Multi-Chain Crypto Wallet and Exchange Platform API
  description: |
    RESTful API for cryptocurrency wallet management, transactions, and exchange operations.
    Supports Bitcoin (BTC), Ethereum (ETH), Solana (SOL), and Stellar (XLM) blockchains.
  version: 1.0.0
  contact:
    name: API Support
    email: api@cryptowallet.example.com
  license:
    name: Proprietary
    url: https://cryptowallet.example.com/license
servers:
  - url: https://api.cryptowallet.example.com/v1
    description: Production server
  - url: https://staging-api.cryptowallet.example.com/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Local development server
security:
  - cookieAuth: []
...
paths:
...
  /wallets:
    get:
      tags: [Wallets]
      summary: List user wallets
      operationId: listWallets
      parameters:
        - name: chain
          in: query
          schema:
            $ref: '#/components/schemas/BlockchainChain'
        - name: status
          in: query
          schema:
            type: string
            enum: [active, archived]
      responses:
        '200':
          description: Wallets retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  wallets:
                    type: array
                    items:
                      $ref: '#/components/schemas/Wallet'
                  total:
                    type: integer

    post:
      tags: [Wallets]
      summary: Create new wallet
      operationId: createWallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWalletRequest'
      responses:
        '201':
          description: Wallet created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
        '400':
          $ref: '#/components/responses/BadRequest'
  /wallets/{id}:
    get:
      tags: [Wallets]
      summary: Get wallet details
      operationId: getWallet
      parameters:
        - $ref: '#/components/parameters/WalletId'
      responses:
        '200':
          description: Wallet retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletDetail'
        '404':
          $ref: '#/components/responses/NotFound'
  /wallets/{id}/balance:
    get:
      tags: [Wallets]
      summary: Get wallet balance
      operationId: getWalletBalance
      parameters:
        - $ref: '#/components/parameters/WalletId'
      responses:
        '200':
          description: Balance retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletBalance'
  /wallets/{id}/transactions:
    get:
      tags: [Wallets]
      summary: Get wallet transactions
      operationId: getWalletTransactions
      parameters:
        - $ref: '#/components/parameters/WalletId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TransactionStatus'
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/TransactionType'
      responses:
        '200':
          description: Transactions retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  total:
                    type: integer
                  pagination:
                    $ref: '#/components/schemas/Pagination'

  # ==================== Transactions ====================
  /transactions:
    get:
      tags: [Transactions]
      summary: List user transactions
      operationId: listTransactions
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: chain
          in: query
          schema:
            $ref: '#/components/schemas/BlockchainChain'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TransactionStatus'
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/TransactionType'
        - name: from_date
          in: query
          schema:
            type: string
            format: date-time
        - name: to_date
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Transactions retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  total:
                    type: integer
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags: [Transactions]
      summary: Create and broadcast transaction
      operationId: createTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionRequest'
      responses:
        '201':
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /transactions/{id}:
    get:
      tags: [Transactions]
      summary: Get transaction details
      operationId: getTransaction
      parameters:
        - $ref: '#/components/parameters/TransactionId'
      responses:
        '200':
          description: Transaction retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDetail'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /rates:
    get:
      tags: [Rates]
      summary: Get current cryptocurrency prices
      operationId: getRates
      parameters:
        - name: symbols
          in: query
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Rates retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  rates:
                    type: array
                    items:
                      $ref: '#/components/schemas/ExchangeRate'
                  last_updated:
                    type: string
                    format: date-time
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: jwt
  parameters:
    WalletId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    TransactionId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
  schemas:
...
    BlockchainChain:
      type: string
      enum: [BTC, ETH, SOL, XLM]
    TransactionStatus:
      type: string
      enum: [pending, confirming, confirmed, failed, cancelled]
    TransactionType:
      type: string
      enum: [send, receive, swap_in, swap_out]
...
    CreateWalletRequest:
      type: object
      required: [chain]
      properties:
        chain:
          $ref: '#/components/schemas/BlockchainChain'
        label:
          type: string
          maxLength: 100
    Wallet:
      type: object
      properties:
        id:
          type: string
          format: uuid
        chain:
          $ref: '#/components/schemas/BlockchainChain'
        address:
          type: string
        label:
          type: string
        balance:
          type: string
          example: "1.5"
        balance_usd:
          type: string
          example: "50000.00"
        status:
          type: string
          enum: [active, archived]
        created_at:
          type: string
          format: date-time
    WalletDetail:
      allOf:
        - $ref: '#/components/schemas/Wallet'
        - type: object
          properties:
            derivation_path:
              type: string
            balance_updated_at:
              type: string
              format: date-time
            transaction_count:
              type: integer
    WalletBalance:
      type: object
      properties:
        wallet_id:
          type: string
          format: uuid
        balance:
          type: string
        balance_usd:
          type: string
        price_usd:
          type: string
        last_updated:
          type: string
          format: date-time

    # ========== Transactions ==========
    CreateTransactionRequest:
      type: object
      required: [wallet_id, to_address, amount]
      properties:
        wallet_id:
          type: string
          format: uuid
        to_address:
          type: string
        amount:
          type: string
        fee_level:
          type: string
          enum: [slow, standard, fast]
          default: standard
        memo:
          type: string
          maxLength: 500
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        wallet_id:
          type: string
          format: uuid
        chain:
          $ref: '#/components/schemas/BlockchainChain'
        tx_hash:
          type: string
        type:
          $ref: '#/components/schemas/TransactionType'
        amount:
          type: string
        fee:
          type: string
        status:
          $ref: '#/components/schemas/TransactionStatus'
        from_address:
          type: string
        to_address:
          type: string
        confirmations:
          type: integer
        created_at:
          type: string
          format: date-time
        confirmed_at:
          type: string
          format: date-time
    TransactionDetail:
      allOf:
        - $ref: '#/components/schemas/Transaction'
        - type: object
          properties:
            block_number:
              type: integer
              format: int64
            explorer_url:
              type: string
              format: uri
            metadata:
              type: object
...
    ExchangeRate:
      type: object
      properties:
        symbol:
          type: string
        price_usd:
          type: string
        price_change_24h:
          type: string
        volume_24h:
          type: string
        market_cap:
          type: string
        last_updated:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
        has_more:
          type: boolean
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...